Moving ChainId enum to constants
