+++ packages/network/src/index.ts
@@ -1,3 +1,4 @@
+export * from './constants'
 export * from './config'
 export * from './json-rpc'
 export * from './json-rpc-provider'
